File Handling Update
Added file handling functionality for customer and account organization

Implemented a system where a customer's folder is created using their TCID or SSN. When an account is created, a corresponding folder is automatically generated inside the customer's folder. This structure helps to organize the project better and allows saving account information and transaction logs under the respective account's folder, making it easier to manage and retrieve data.